+++ src/auth/AuthProtectedRoute.tsx
@@ -1,0 +1,41 @@
+import useLogin from "auth/useLogin";
+import useUser from "components/context/UserContext";
+import { PageLoader } from "components/LoadingIndicator";
+import React from "react";
+
+const AuthProtectedRoute: React.FC = ({ children }) => {
+  const { isLoading, isAuthenticated } = useUser();
+  const { initLogin } = useLogin();
+
+  React.useEffect(() => {
+    if (!isLoading && !isAuthenticated) {
+      initLogin();
+    }
+  }, [initLogin, isLoading, isAuthenticated]);
+
+  if (isAuthenticated) {
+    return <>{children}</>;
+  }
+  if (isLoading) {
+    return <PageLoader />;
+  }
+  return <Unauthorized />;
+};
+
+export default AuthProtectedRoute;
+
+const Unauthorized = () => {
+  return (
+    <div
+      sx={{
+        flex: 1,
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        flexDirection: "column"
+      }}
+    >
+      <h4>You need to be signed in to view this page.</h4>
+    </div>
+  );
+};
