+++ src/components/MyBooks.tsx
@@ -10,7 +10,7 @@ import PageTitle from "./PageTitle";
 import SignOut from "./SignOut";
 import useUser from "components/context/UserContext";
 import { PageLoader } from "components/LoadingIndicator";
-import useLogin from "auth/useLogin";
+import AuthProtectedRoute from "auth/AuthProtectedRoute";
 
 const availableUntil = (book: AnyBook) =>
   book.availability?.until ? new Date(book.availability.until) : "NaN";
@@ -39,32 +39,13 @@ function compareTitles(a: AnyBook, b: AnyBook): 0 | -1 | 1 {
   return -1;
 }
 
-const AuthProtected: React.FC = ({ children }) => {
-  const { isLoading, isAuthenticated } = useUser();
-  const { initLogin } = useLogin();
-
-  React.useEffect(() => {
-    if (!isLoading && !isAuthenticated) {
-      initLogin();
-    }
-  }, [initLogin, isLoading, isAuthenticated]);
-
-  if (isAuthenticated) {
-    return children;
-  }
-  if (isLoading) {
-    return <PageLoader />;
-  }
-  return <Unauthorized />;
-};
-
 export const MyBooks: React.FC = () => {
-  const { isAuthenticated, loans, isLoading } = useUser();
+  const { loans, isLoading } = useUser();
   const sortedBooks = loans ? sortBooksByLoanExpirationDate(loans) : [];
   const noBooks = sortedBooks.length === 0;
 
   return (
-    <AuthProtected>
+    <AuthProtectedRoute>
       <div sx={{ flex: 1, pb: 4 }}>
         <Head>
           <title>My Books</title>
@@ -74,15 +55,13 @@ export const MyBooks: React.FC = () => {
         <PageTitle>My Books</PageTitle>
         {noBooks && isLoading ? (
           <PageLoader />
-        ) : isAuthenticated && noBooks ? (
+        ) : noBooks ? (
           <Empty />
-        ) : isAuthenticated ? (
-          <LoansContent books={sortedBooks} />
         ) : (
-          <Unauthorized />
+          <LoansContent books={sortedBooks} />
         )}
       </div>
-    </AuthProtected>
+    </AuthProtectedRoute>
   );
 };
 
@@ -91,25 +70,6 @@ const LoansContent: React.FC<{ books: AnyBook[] }> = ({ books }) => {
     <React.Fragment>
       <BookList books={books} />
     </React.Fragment>
-  );
-};
-
-const Unauthorized = () => {
-  return (
-    <div
-      sx={{
-        flex: 1,
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        flexDirection: "column"
-      }}
-    >
-      <Head>
-        <title>My Books</title>
-      </Head>
-      <h4>You need to be signed in to view this page.</h4>
-    </div>
   );
 };
